requirements.txt
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
APScheduler==3.6.0
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
blinker==1.4
bokeh==1.0.4
boto==2.49.0
boto3==1.9.111
botocore==1.12.111
Bottleneck==1.2.1
bz2file==0.98
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
colorama==0.4.1
conda==4.6.9
conda-build==3.17.8
conda-verify==3.1.1
contextlib2==0.5.5
cryptography==2.5
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
dask==1.1.3
datashape==0.5.4
decorator==4.3.2
defusedxml==0.5.0
distributed==1.26.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
findspark==1.0.0
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
gensim==3.4.0
gevent==1.3.7
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.13
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.0.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.13
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.3
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.26.0
locket==0.2.0
lxml==4.3.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
more-itertools==4.3.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.2.5
nose==1.3.7
notebook==5.7.4
numba==0.41.0
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
oauthlib==3.0.1
odo==0.5.1
olefile==0.46
openpyxl==2.6.1
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pkginfo==1.5.0.1
pluggy==0.9.0
plumbum==1.6.7
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.0
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.1
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.3.0
pytest-arraydiff==0.3
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-crfsuite==0.9.6
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
PyYAML==3.13
pyzmq==18.0.1
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.6.0
requests==2.21.0
requests-oauthlib==1.2.0
rope==0.10.7
rpyc==4.0.2
ruamel-yaml==0.15.71
s3transfer==0.2.0
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.4.2
SQLAlchemy==1.3.0
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
twython==3.7.0
typed-ast==1.3.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.3
xlwt==1.3.0
zict==0.1.3
zipp==0.3.3
